# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    FeedPostType:
        type: string
        x-go-package: github.com/ogniloud/madr/internal/social/models
    FeedResponse:
        properties:
            feed:
                items:
                    $ref: '#/definitions/Post'
                type: array
                x-go-name: Feed
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    FollowersResponse:
        properties:
            user_info:
                items:
                    $ref: '#/definitions/UserInfo'
                type: array
                x-go-name: UserInfo
        title: FollowersResponse contains ids and names of the followers.
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
    FollowingSubscribedData:
        properties:
            author_id:
                description: AuthorId is a unique identifier of the author.
                example: 1
                format: int64
                type: integer
                x-go-name: AuthorId
            author_name:
                description: AuthorName is a name of the author.
                example: Dima Krasnov
                type: string
                x-go-name: AuthorName
            follower_id:
                description: FollowerId is a unique identifier of the follower.
                example: 1
                format: int64
                type: integer
                x-go-name: FollowerId
            follower_name:
                description: FollowerName is a name of the follower.
                example: John Doe
                type: string
                x-go-name: FollowerName
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
    FollowingsResponse:
        properties:
            user_info:
                items:
                    $ref: '#/definitions/UserInfo'
                type: array
                x-go-name: UserInfo
        title: FollowingsResponse contains ids and names of the followings.
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
    GenericError:
        properties:
            message:
                description: The error message.
                example: Very useful error message
                type: string
                x-go-name: Message
        title: GenericError is a generic error message returned by a server.
        type: object
        x-go-package: github.com/ogniloud/madr/internal/ioutil
...
    InviteData:
        properties:
            group_id:
                description: GroupId is a unique identifier of the group.
                example: 1
                format: int64
                type: integer
                x-go-name: GroupId
            group_name:
                description: GroupName is a name of the group.
                example: vodka lovers
                type: string
                x-go-name: GroupName
            invitee_id:
                description: InviteeId is a unique identifier of the user that received the invitation.
                example: 1
                format: int64
                type: integer
                x-go-name: InviteeId
            invitee_name:
                description: InviteeName is a name of the user that received the invitation.
                example: John Doe
                type: string
                x-go-name: InviteeName
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    Post:
        properties:
            following_subscribed_data:
                $ref: '#/definitions/FollowingSubscribedData'
            invite_data:
                $ref: '#/definitions/InviteData'
            shared_from_following_data:
                $ref: '#/definitions/SharedFromFollowingData'
            shared_from_group_data:
                $ref: '#/definitions/SharedFromGroupData'
            type:
                $ref: '#/definitions/FeedPostType'
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    SearchUserResponse:
        description: |-
            SearchUserResponse is a struct that defines the response body for
            searching users by name.
        properties:
            users:
                items:
                    $ref: '#/definitions/UserInfo'
                type: array
                x-go-name: Users
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
    SharedFromFollowingData:
        properties:
            author_id:
                description: AuthorId is a unique identifier of the author.
                example: 1
                format: int64
                type: integer
                x-go-name: AuthorId
            author_name:
                description: AuthorName is a name of the author.
                example: John Doe
                type: string
                x-go-name: AuthorName
            deck_id:
                description: DeckId is a unique identifier of the deck.
                example: 1
                format: int64
                type: integer
                x-go-name: DeckId
            deck_name:
                description: DeckName is a name of the deck.
                example: russian curse words
                type: string
                x-go-name: DeckName
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
    SharedFromGroupData:
        properties:
            deck_id:
                description: DeckId is a unique identifier of the deck.
                example: 1
                format: int64
                type: integer
                x-go-name: DeckId
            deck_name:
                description: DeckName is a name of the deck.
                example: Bangladesh famous words
                type: string
                x-go-name: DeckName
            group_id:
                description: GroupId is a unique identifier of the group.
                example: 1
                format: int64
                type: integer
                x-go-name: GroupId
            group_name:
                description: GroupName is a name of the group.
                example: vodka lovers
                type: string
                x-go-name: GroupName
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    UserInfo:
        properties:
            Email:
                type: string
            ID:
                format: int64
                type: integer
            Username:
                type: string
        title: UserInfo is a struct that defines the user info model.
        type: object
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    feedRequest:
        properties:
            user_id:
                format: int64
                type: integer
                x-go-name: UserId
        title: FeedRequest is a struct that defines the request body for getting a user's feed.
        type: object
        x-go-name: FeedRequest
        x-go-package: github.com/ogniloud/madr/internal/social/models
    followRequest:
        description: FollowRequest contains an id of the user that following to author
        properties:
            author_id:
                description: AuthorId is an ID of the supplier of content.
                example: 189
                format: int64
                type: integer
                x-go-name: AuthorId
            follower_id:
                description: FollowerId is an ID of the user following the author.
                example: 189
                format: int64
                type: integer
                x-go-name: FollowerId
        required:
            - follower_id
            - author_id
        type: object
        x-go-name: FollowRequest
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
    getFollowersRequest:
        properties:
            user_id:
                description: UserId is an ID of the user in a storage.
                example: 189
                format: int64
                type: integer
                x-go-name: UserId
        required:
            - user_id
        title: FollowersRequest contains an id of the user to get their followers.
        type: object
        x-go-name: FollowersRequest
        x-go-package: github.com/ogniloud/madr/internal/social/models
    getFollowingsRequest:
        properties:
            user_id:
                description: UserId is an ID of the user in a storage.
                example: 189
                format: int64
                type: integer
                x-go-name: UserId
        required:
            - user_id
        title: FollowingsRequest contains an id of the user to get their followings.
        type: object
        x-go-name: FollowingsRequest
        x-go-package: github.com/ogniloud/madr/internal/social/models
...
info:
    contact:
        email: pelageech@mail.ru
        name: Artyom Blaginin
    description: Documentation for MADR API.
    title: describes models for application.
    version: 0.0.1
paths:
...
    /api/social/feed:
        post:
            consumes:
                - application/json
            operationId: Feed
            parameters:
                - description: request body
                  in: body
                  name: request
                  required: true
                  schema:
                    $ref: '#/definitions/feedRequest'
                    description: request body
                    type: object
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "200":
                    $ref: '#/responses/feedOkResponse'
                "400":
                    $ref: '#/responses/feedBadRequestResponse'
                "500":
                    $ref: '#/responses/feedInternalServerErrorResponse'
            summary: Returns user's feed.
    /api/social/follow:
        post:
            consumes:
                - application/json
            operationId: follow
            parameters:
                - description: request body
                  in: body
                  name: request
                  required: true
                  schema:
                    $ref: '#/definitions/followRequest'
                    description: request body
                    type: object
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "204":
                    $ref: '#/responses/followNoContentResponse'
                "400":
                    $ref: '#/responses/followBadRequestResponse'
                "500":
                    $ref: '#/responses/followInternalServerErrorResponse'
            summary: Follows the user.
    /api/social/followers:
        post:
            consumes:
                - application/json
            operationId: getFollowers
            parameters:
                - description: request body
                  in: body
                  name: request
                  required: true
                  schema:
                    $ref: '#/definitions/getFollowersRequest'
                    description: request body
                    type: object
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "200":
                    $ref: '#/responses/getFollowersOkResponse'
                "400":
                    $ref: '#/responses/getFollowersBadRequestResponse'
                "500":
                    $ref: '#/responses/getFollowersInternalServerErrorResponse'
            summary: Returns a list of followers of the user.
    /api/social/followings:
        post:
            consumes:
                - application/json
            operationId: getFollowings
            parameters:
                - description: request body
                  in: body
                  name: request
                  required: true
                  schema:
                    $ref: '#/definitions/getFollowingsRequest'
                    description: request body
                    type: object
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "200":
                    $ref: '#/responses/getFollowingsOkResponse'
                "400":
                    $ref: '#/responses/getFollowingsBadRequestResponse'
                "500":
                    $ref: '#/responses/getFollowingsInternalServerErrorResponse'
            summary: Returns a list of followings of the user.
    /api/social/search:
        get:
            consumes:
                - application/json
            operationId: search
            parameters:
                - description: search query
                  in: query
                  name: q
                  required: true
                  type: string
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "200":
                    $ref: '#/responses/searchUserOkResponse'
                "500":
                    $ref: '#/responses/searchUserInternalServerErrorResponse'
            summary: Returns a list of users by name.
    /api/social/unfollow:
        post:
            consumes:
                - application/json
            operationId: unfollow
            parameters:
                - description: request body
                  in: body
                  name: request
                  required: true
                  schema:
                    $ref: '#/definitions/followRequest'
                    description: request body
                    type: object
            produces:
                - application/json
                - 'Scheme: http'
            responses:
                "204":
                    $ref: '#/responses/unfollowNoContentResponse'
                "400":
                    $ref: '#/responses/unfollowBadRequestResponse'
                "500":
                    $ref: '#/responses/unfollowInternalServerErrorResponse'
            summary: Unfollows the user.
...
produces:
    - application/json
responses:
...
    feedBadRequestResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    feedInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    feedOkResponse:
        description: ""
        schema:
            $ref: '#/definitions/FeedResponse'
    followBadRequestResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    followInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    followNoContentResponse:
        description: ""
...
    getFollowersBadRequestResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    getFollowersInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    getFollowersOkResponse:
        description: ""
        schema:
            $ref: '#/definitions/FollowersResponse'
    getFollowingsBadRequestResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    getFollowingsInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    getFollowingsOkResponse:
        description: ""
        schema:
            $ref: '#/definitions/FollowingsResponse'
...
    searchUserInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    searchUserOkResponse:
        description: ""
        schema:
            $ref: '#/definitions/SearchUserResponse'
...
    unfollowBadRequestResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    unfollowInternalServerErrorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    unfollowNoContentResponse:
        description: ""
schemes:
    - http
swagger: "2.0"